+++ CHANGELOG.md
@@ -79,7 +79,8 @@ Initial version
   * reordered UI elements to save horizontal space
 
 ## Version 0.2.7 (xx.xx.xxxx)
-* add read_custom_field() function
+* add readCustomField() function
 * fixed serial2Df (use correct path & properly handle serial logs which contain carriage returns)
+* added functions to read and write symbol values in ELF files (binary patching)
 * xml config
   * no longer add `remoteIp` field in xml if no value specified (to improve serial logging performance)